Change to my phonetic alphabet
It's pretty entertaining to yell "alpha bravo whisky tango" at your
computer, but the NATO phonetic alphabet was designed to be
understandable over fuzzy radios in the middle of combat. I'm a
centimeter away from a nice mic, so I can get away with shorter words
that are easier to say.
This alphabet came from somebody else originally, and I've been
tinkering with it for 3 years. It fits my voice really well, but I have
no idea how well it generalizes to other people.

This alphabet + vim provides 90% of the functionality I need for
dictating source code.